requirements.txt
requests
python-dotenv
plyer
colorama